direct translation of renderer.js into MainController.tsx
